add text for not yet implemented parts
